refactor: move salon subdirs into component dirs
- EmotionSelector: selected_emotions
- FaqList: collapse, left_right, search_hint
- KanbanThread: classic_layout, waterfall_layout
- Share: modal_panel
