+++ frontend/core/unit/EmotionSelector/SelectedEmotions/EmotionIcon.tsx
@@ -3,7 +3,7 @@ import { type FC, memo } from 'react'
 import Img from '~/Img'
 import type { TEmotionType } from '~/spec'
 
-import useSalon from '../salon/selected_emotions/emotion_icon'
+import useSalon from './salon/emotion_icon'
 
 type TProps = {
   name: TEmotionType
