+++ frontend/core/unit/EmotionSelector/SelectedEmotions/salon/emotion_icon.ts
