+++ frontend/core/unit/EmotionSelector/SelectedEmotions/salon/emotion_unit.ts
