+++ frontend/core/unit/EmotionSelector/SelectedEmotions/salon/users_panel.ts
