+++ frontend/core/unit/FaqList/Collapse/salon/banner.ts
