+++ frontend/core/unit/FaqList/Collapse/salon/footer.ts
