+++ frontend/core/unit/FaqList/Collapse/salon/section.ts
