+++ frontend/core/unit/FaqList/SearchHint/salon/cat_section.ts
