+++ frontend/core/unit/FaqList/SearchHint/salon/index.ts
