+++ frontend/core/unit/KanbanThread/ClassicLayout/salon/actions.ts
