+++ frontend/core/unit/KanbanThread/ClassicLayout/salon/columns.ts
