+++ frontend/core/unit/KanbanThread/ClassicLayout/salon/index.ts
