+++ frontend/core/unit/KanbanThread/WaterfallLayout/salon/sections.ts
