+++ frontend/core/unit/Share/ModalPanel/LinkBoard.tsx
@@ -3,7 +3,7 @@ import { type FC, memo, useState } from 'react'
 import CopyButton from '~/widgets/Buttons/CopyButton'
 import Input from '~/widgets/Input'
 
-import useSalon, { cn } from '../salon/modal_panel/link_board'
+import useSalon, { cn } from './salon/link_board'
 import type { TLinksData } from '../spec'
 
 type TProps = {
