+++ frontend/core/unit/Share/ModalPanel/index.tsx
@@ -8,7 +8,7 @@ import type { FC } from 'react'
 import type { TArticle } from '~/spec'
 import Modal from '~/widgets/Modal'
 
-import useSalon, { cn } from '../salon/modal_panel'
+import useSalon, { cn } from './salon'
 import type { TLinksData } from '../spec'
 import InfoPanel from './InfoPanel'
 import Platforms from './Platforms'
