+++ frontend/core/unit/Share/ModalPanel/salon/link_board.ts
