+++ frontend/core/unit/Share/ModalPanel/salon/platform.ts
