+++ frontend/core/unit/Share/ModalPanel/salon/wechat_board.ts
